TA: OGE-94 build libcull as C++ code

+++ source/libs/cull/cull_multitype.cc
@@ -190,7 +190,7 @@ int lGetPosViaElem(const lListElem *element, int name, int do_abort)
 ******************************************************************************/
 const char *lMt2Str(int mt) 
 {
-   if (mt >= 0 && mt < sizeof(multitypes)/sizeof(char*)) {
+   if (mt >= 0 && mt < (int)(sizeof(multitypes)/sizeof(char*))) {
       return multitypes[mt];
    } else {
       return "unknown multitype";
